requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyascii==0.3.2
async-timeout==5.0.1
attrs==25.1.0
audioread==3.0.1
babel==2.17.0
bangla==0.0.2
beautifulsoup4==4.13.3
blinker==1.9.0
blis==1.2.0
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
comtypes==1.4.10
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.11
Cython==3.0.12
dateparser==1.1.8
decorator==5.2.1
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
filelock==3.17.0
Flask==3.1.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
g2pkk==0.1.2
grpcio==1.70.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
hangul-romanize==0.1.0
huggingface-hub==0.29.1
idna==3.10
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.5
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.6.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.12
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numba==0.60.0
numpy==1.22.0
openai-whisper==20240930
packaging==24.2
pandas==1.5.3
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pynndescent==0.5.13
pyparsing==3.2.1
pyperclip==1.9.0
pypinyin==0.53.0
pypiwin32==223
PyRect==0.2.0
pysbd==0.3.4
PyScreeze==1.0.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytweening==1.2.0
pytz==2025.1
pywhatkit==5.4
pywin32==309
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.11.4
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
SudachiDict-core==20250129
SudachiPy==0.6.10
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.5.0
torchaudio==2.5.0
tqdm==4.67.1
trainer==0.0.36
transformers==4.49.0
TTS==0.22.0
typeguard==4.4.2
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
umap-learn==0.5.7
Unidecode==1.3.8
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
webrtcvad==2.0.10
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.17.2
yarl==1.18.3
